+++ src/app/table/notifier/aria-notifier.component.ts
@@ -4,11 +4,14 @@ import { Subscription } from 'rxjs/Subscription';
 
 @Component({
   selector: 'shark-table-aria-notifier',
-  template: `<div class="notification-area screen-reader" aria-atomic="true" aria-live="polite">{{ status }}</div>`
+  template: `<div id="shark-table-aria-notifier-{{ tableId }}" class="notification-area screen-reader" aria-atomic="true" aria-live="polite">{{ status }}</div>`
 })
 export class AriaNotifierComponent implements OnChanges, OnDestroy {
 
   status: string;
+
+  @Input()
+  tableId: string;
 
   @Input()
   notifierService: NotifierService;